resolver: defer the type-argument dispatch gate to the identity work
Remove the closure proof and the type-argument filter from the C#
interface-dispatch pass, returning it to the ungated fan-out. Five
review rounds showed the gate approximating evidence the graph cannot
yet supply - unique type node identity, per-construction implements
edges, resolved receiver bindings - and each approximation kept opening
new recall regressions. The extraction-side evidence (type-argument
stamps, variance stamps, canonical spellings, the shadow index) stays
and keeps maturing; the gate returns when a receiver can carry a
resolved field node ID and an implements edge its own construction.

The gate test families go with it, including the dispatch benchmark
that exceeded the CI bench-job timeout. The recall pins that guard
extraction-side behaviour end to end stay, with their shared helpers
moved out of the removed suite.
