--- internal/resolver/csharp_iface_dispatch_binding_forms_test.go
@@ -1,258 +1,0 @@
-package resolver
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/zzet/gortex/internal/graph"
-)
-
-// bindMemberCallAtLine mirrors the enrichment/LSP binding for a member
-// call whose extraction companion carries NO receiver evidence - the
-// exact state a shadow-refused site is in. bindFieldReceiverCall finds
-// the companion by its receiver_name stamp, which such a site does not
-// have; here the companion is found by member name alone.
-func bindMemberCallAtLine(t *testing.T, g graph.Store, callerID, memberName, target string) {
-	t.Helper()
-	var companion *graph.Edge
-	for _, e := range g.GetOutEdges(callerID) {
-		if e != nil && e.Kind == graph.EdgeCalls && e.To == "unresolved::*."+memberName {
-			companion = e
-			break
-		}
-	}
-	require.NotNil(t, companion, "fixture: the extraction must leave an unresolved companion for the member call")
-	g.AddEdge(&graph.Edge{
-		From: callerID, To: target, Kind: graph.EdgeCalls,
-		FilePath: companion.FilePath, Line: companion.Line,
-		Origin: graph.OriginASTResolved, Confidence: 0.95,
-	})
-}
-
-// The shadow indexes were fed by exactly two binding forms - the
-// local_declaration_statement capture and emitted KindParam nodes. C#
-// has more: foreach variables, declaration patterns (`o is T x`),
-// out-var declaration expressions, lambda parameters, and the
-// parenthesized `using (var x = ...)` resource. A name bound by any of
-// them was invisible to the refusal, so when it coincided with a field
-// name the site stamped receiver_name, the field-identifier emitter
-// minted the read-edge evidence off the SAME index, and the gate
-// filtered on the field's closure - a receiver the call site does not
-// have. Both layers of the two-layer guard failed together, because a
-// two-layer guard whose layers share one index is one layer.
-//
-// Every fixture binds `_box` (an IBox<Widget>) over a call site while
-// the enclosing type declares `IBox<Crate> _box`. The receiver is the
-// bound name, not the field, so no closure is provable and the site
-// must keep the full fan-out.
-func TestResolveCSharpInterfaceDispatch_BindingFormsShadowFieldReceivers(t *testing.T) {
-	for _, tc := range []struct {
-		name   string
-		caller string
-		body   string
-	}{
-		{"foreach variable", "Flow.Sum", `
-        public int Sum(IBox<Widget>[] all) {
-            int t = 0;
-            foreach (var _box in all) { t += _box.Get(7); }
-            return t;
-        }`},
-		{"declaration pattern", "Flow.Check", `
-        public int Check(object o) {
-            if (o is IBox<Widget> _box) { return _box.Get(7); }
-            return 0;
-        }`},
-		{"out var", "Flow.Pull", `
-        private bool TryMake(out IBox<Widget> made) { made = null; return false; }
-        public int Pull() {
-            if (TryMake(out var _box)) { return _box.Get(7); }
-            return 0;
-        }`},
-		{"lambda parameter", "Flow.Total", `
-        public int Total(IBox<Widget>[] all) {
-            return all.Sum(_box => _box.Get(7));
-        }`},
-		{"parenthesized using", "Flow.Use", `
-        private IBox<Widget> Make(IBox<Widget> s) { return s; }
-        public int Use(IBox<Widget> src) {
-            using (var _box = Make(src)) { return _box.Get(7); }
-        }`},
-		{"catch variable", "Flow.Guard", `
-        public int Guard() {
-            try { return 0; }
-            catch (BoxError _box) { return _box.Get(7); }
-        }`},
-		{"for initializer", "Flow.Loop", `
-        private IBox<Widget> Make(IBox<Widget> s) { return s; }
-        public int Loop(IBox<Widget> src) {
-            for (var _box = Make(src); _box != null; ) { return _box.Get(7); }
-            return 0;
-        }`},
-		{"anonymous method parameter", "Flow.Old", `
-        public int Old(IBox<Widget> src) {
-            System.Func<IBox<Widget>, int> f = delegate(IBox<Widget> _box) { return _box.Get(7); };
-            return f(src);
-        }`},
-		{"query from variable", "Flow.Query", `
-        public int Query(IBox<Widget>[] all) {
-            return System.Linq.Enumerable.Sum(from _box in all select _box.Get(7));
-        }`},
-		{"query let variable", "Flow.Bind", `
-        public int Bind(IBox<Widget>[] all) {
-            return System.Linq.Enumerable.Sum(from w in all let _box = w select _box.Get(7));
-        }`},
-		{"local function parameter", "Flow.Host", `
-        public int Host() {
-            int Inner(IBox<Widget> _box) { return _box.Get(7); }
-            return Inner(null);
-        }`},
-		{"recursive pattern designation", "Flow.Probe", `
-        public int Probe(object o) {
-            if (o is IBox<Widget> { } _box) { return _box.Get(7); }
-            return 0;
-        }`},
-		{"parenthesized designation", "Flow.Split", `
-        public int Split(object o) {
-            if (o is var (_box, n)) { return _box.Get(7); }
-            return 0;
-        }`},
-		{"deconstruction declaration", "Flow.Take", `
-        public int Take((IBox<Widget>, int) t) {
-            var (_box, n) = t;
-            return _box.Get(7);
-        }`},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			g := buildCSharpResolverGraph(t, map[string]string{
-				"F.cs": `namespace App {
-    public class Crate { }
-    public class Widget { }
-    public interface IBox<T> { int Get(int id); }
-    public class CrateBox : IBox<Crate> { public int Get(int id) { return 1; } }
-    public class WidgetBox : IBox<Widget> { public int Get(int id) { return 2; } }
-    public class BoxError : System.Exception {
-        public int Get(int id) { return 0; }
-    }
-    public class Flow {
-        private readonly IBox<Crate> _box;
-        public Flow(IBox<Crate> b) { _box = b; }
-` + tc.body + `
-    }
-}`,
-			})
-			New(g).ResolveAll()
-
-			callerID := "F.cs::" + tc.caller
-			bindMemberCallAtLine(t, g, callerID, "Get", "F.cs::IBox.Get")
-			ResolveCSharpInterfaceDispatch(g)
-
-			assert.ElementsMatch(t, []string{
-				"F.cs::CrateBox.Get",
-				"F.cs::WidgetBox.Get",
-			}, dispatchTargets(g, callerID),
-				"the receiver is the bound name, not the field - no closure is provable, so the full fan-out stays")
-		})
-	}
-}
-
-// A receiverless call is a call on `this`, and it can share a line with
-// a field-receiver call of the same member name:
-//
-//	return Get(1) + _widgets.Get(2);
-//
-// The bare call's bound edge carries no receiver meta, so the receiver
-// join falls back to the only `unresolved::*.Get` companion on the line
-// - the SIBLING's - and adopts its Widget closure for a call whose real
-// receiver is a Flow (an IBox<Crate>). The actual callee is dropped and
-// the type-impossible implementor is the only survivor.
-//
-// The site-ambiguity index skipped receiverless calls entirely, so the
-// steal went unmarked. An empty receiver differing from `_widgets` is
-// exactly as disqualifying as a different spelled one.
-func TestResolveCSharpInterfaceDispatch_ImplicitThisNeverStealsSiblingReceiver(t *testing.T) {
-	g := buildCSharpResolverGraph(t, map[string]string{
-		"F.cs": `namespace App {
-    public class Crate { }
-    public class Widget { }
-    public interface IBox<T> { int Get(int id); }
-    public class WidgetBox : IBox<Widget> { public int Get(int id) { return 2; } }
-    public class Mid : IBox<Crate> { public int Get(int id) { return 1; } }
-    public class Flow : Mid {
-        private readonly IBox<Widget> _widgets;
-        public Flow(IBox<Widget> w) { _widgets = w; }
-        public int Pull() { return Get(1) + _widgets.Get(2); }
-    }
-}`,
-	})
-	New(g).ResolveAll()
-
-	// The bare Get(1) binds through the interface family like any
-	// member call the enrichment tier resolves; its bound edge carries
-	// no receiver evidence of its own.
-	const callerID = "F.cs::Flow.Pull"
-	bindMemberCallAtLine(t, g, callerID, "Get", "F.cs::IBox.Get")
-	ResolveCSharpInterfaceDispatch(g)
-
-	// With the line marked ambiguous nothing at the site may filter, so
-	// the fan-out keeps every implementor - the actual callee Mid.Get
-	// included.
-	assert.ElementsMatch(t, []string{
-		"F.cs::Mid.Get",
-		"F.cs::WidgetBox.Get",
-	}, dispatchTargets(g, callerID),
-		"a receiverless implicit-this call marks its line ambiguous rather than borrowing the sibling's receiver")
-}
-
-// Two members declared on one source line used to tie for the innermost
-// func range with an extraction-order tie-break, handing A's call to B.
-// Byte-interval attribution (round-5 finding 4) resolves the tie: the
-// call's offset lies inside A's extent, so A owns it. What this test
-// still pins is the EVIDENCE story on such a line: the receiver spells
-// A's shadowing parameter, receiver_name is refused, the line stays
-// marked ambiguous for the same-name-site index - so nothing at the
-// site may filter and the fan-out keeps every implementor.
-func TestResolveCSharpInterfaceDispatch_TwoMembersOnOneLineRefuseReceiverEvidence(t *testing.T) {
-	g := buildCSharpResolverGraph(t, map[string]string{
-		"F.cs": `namespace App {
-    public class Crate { }
-    public class Widget { }
-    public interface IBox<T> { int Get(int id); }
-    public class CrateBox : IBox<Crate> { public int Get(int id) { return 1; } }
-    public class WidgetBox : IBox<Widget> { public int Get(int id) { return 2; } }
-    public class Flow {
-        private readonly IBox<Crate> _store;
-        public Flow(IBox<Crate> s) { _store = s; }
-        public int B() { return 0; } public int A(IBox<Widget> _store, int id) { return _store.Get(id); }
-    }
-}`,
-	})
-	New(g).ResolveAll()
-
-	// Byte attribution hands the call to its real owner, A; B owns no
-	// call at all.
-	assert.Empty(t, callsFrom(g, "F.cs::Flow.B"),
-		"B's body is `return 0;` - byte attribution leaves it no call")
-
-	const callerID = "F.cs::Flow.A"
-	bindMemberCallAtLine(t, g, callerID, "Get", "F.cs::IBox.Get")
-	ResolveCSharpInterfaceDispatch(g)
-
-	assert.ElementsMatch(t, []string{
-		"F.cs::CrateBox.Get",
-		"F.cs::WidgetBox.Get",
-	}, dispatchTargets(g, callerID),
-		"the receiver spells A's shadowing parameter - no usable receiver evidence, so the full fan-out stays")
-}
-
-// callsFrom returns the EdgeCalls targets leaving fromID.
-func callsFrom(g graph.Store, fromID string) []string {
-	var out []string
-	for _, e := range g.GetOutEdges(fromID) {
-		if e != nil && e.Kind == graph.EdgeCalls {
-			out = append(out, e.To)
-		}
-	}
-	return out
-}
